RPN-1080 Fix bestBid/Sell evaluation
