+++ nft-order/listener/src/main/kotlin/com/rarible/protocol/nftorder/listener/evaluator/BestOrderEvaluator.kt
@@ -81,8 +81,8 @@ class BestOrderEvaluator(
         }
 
         logger.info(
-            "Evaluated {} for {} [{}] (current = [{}], updated = [{}])",
-            name, type, id, current.hash, updated.hash
+            "Evaluated {} for {} [{}] (current = [{}], updated = [{}], best = [{}])",
+            name, type, id, current.hash, updated.hash, bestOrder.hash
         )
         return bestOrder
     }
