+++ nft-order/listener/src/main/kotlin/com/rarible/protocol/nftorder/listener/event/KafkaNftOrderItemEventListener.kt
@@ -7,6 +7,7 @@ import com.rarible.protocol.dto.NftOrderItemEventDto
 import com.rarible.protocol.dto.NftOrderItemEventTopicProvider
 import com.rarible.protocol.nftorder.core.event.ItemEvent
 import com.rarible.protocol.nftorder.core.event.ItemEventListener
+import org.slf4j.LoggerFactory
 import org.springframework.core.convert.ConversionService
 import org.springframework.stereotype.Component
 
@@ -16,9 +17,12 @@ class KafkaNftOrderItemEventListener(
     private val eventsProducer: RaribleKafkaProducer<NftOrderItemEventDto>
 ) : ItemEventListener {
 
+    private val logger = LoggerFactory.getLogger(javaClass)
+
     private val eventsHeaders = mapOf(
         "protocol.nft-order.event.version" to NftOrderItemEventTopicProvider.VERSION
     )
+
 
     override suspend fun onEvent(event: ItemEvent) {
         val dto = conversionService.convert<NftOrderItemEventDto>(event)
@@ -29,5 +33,6 @@ class KafkaNftOrderItemEventListener(
             headers = eventsHeaders
         )
         eventsProducer.send(message).ensureSuccess()
+        logger.info("Item Event sent: {}", dto)
     }
 }
