+++ nft-order/listener/src/main/kotlin/com/rarible/protocol/nftorder/listener/service/ItemEventService.kt
@@ -106,13 +106,15 @@ class ItemEventService(
             val updated = orderUpdateAction(item)
             if (item != updated) {
                 if (EnrichmentDataVerifier.isItemNotEmpty(updated)) {
-                    itemService.save(updated)
-                    notify(ItemEventUpdate(updated))
+                    val saved = updateItem(item, updated)
+                    notify(ItemEventUpdate(saved))
                 } else if (!fetchedItem.isFetched) {
                     itemService.delete(itemId)
                     logger.info("Deleted Item [{}] without enrichment data", itemId)
                     notify(ItemEventUpdate(updated))
                 }
+            } else {
+                logger.info("Item [{}] not changed after order updated, event won't be published")
             }
         }
     }
